+++ data/src/main/java/com/teamagam/gimelgimel/data/message/poller/MessageLongPoller.java
@@ -1,7 +1,6 @@
 package com.teamagam.gimelgimel.data.message.poller;
 
 import com.teamagam.gimelgimel.data.config.Constants;
-import com.teamagam.gimelgimel.data.message.adapters.ChatAlert;
 import com.teamagam.gimelgimel.data.message.adapters.ServerDataMapper;
 import com.teamagam.gimelgimel.data.message.entity.MessageAlertData;
 import com.teamagam.gimelgimel.data.message.entity.MessageData;
@@ -148,10 +147,7 @@ public class MessageLongPoller implements IMessagePoller {
 
     @Override
     public void visit(MessageAlertData message) {
-      ChatAlert chatAlert = mServerDataMapper.transform(message);
-
-      mProcessor.process(chatAlert.getMessage());
-      mProcessor.process(chatAlert.getAlert(), chatAlert.getMessage().getMessageId());
+      mProcessor.process(mServerDataMapper.transform(message));
     }
   }
 }
