+++ domain/src/main/java/com/teamagam/gimelgimel/domain/messages/poller/IPolledMessagesProcessor.java
@@ -1,6 +1,5 @@
 package com.teamagam.gimelgimel.domain.messages.poller;
 
-import com.teamagam.gimelgimel.domain.alerts.entity.Alert;
 import com.teamagam.gimelgimel.domain.layers.entitiy.VectorLayer;
 import com.teamagam.gimelgimel.domain.location.entity.UserLocation;
 import com.teamagam.gimelgimel.domain.messages.entity.ChatMessage;
@@ -17,6 +16,4 @@ public interface IPolledMessagesProcessor {
   void process(VectorLayer vectorLayer);
 
   void process(UserLocation userLocation);
-
-  void process(Alert alert, String messageId);
 }
